Added new monorepo package `target-js`
