+++ kipper/target-js/.badge-config
@@ -1,0 +1,10 @@
+{
+  "coverage_file_path": "../../coverage/coverage-summary.json",
+  "badges": {
+    "coverage": {
+      "style": "flat",
+      "logoColor": "white",
+      "color": "blue"
+    }
+  }
+}
