+++ kipper/target-js/.eslintignore
@@ -1,0 +1,4 @@
+lib
+img
+node_modules
+kipper-standalone.*
