+++ kipper/target-js/DEVELOPMENT.md
@@ -1,0 +1,55 @@
+# Developer's Guide
+
+We recommend Visual Studio Code or Web-Storm for developing on Kipper.
+
+## General Setup
+
+### Install the TypeScript Compiler and Node.js
+
+- [Download page of TypeScript](https://www.typescriptlang.org/download)
+- [Download page of Node.js](https://nodejs.org/en/download/)
+
+### VSCode setup
+
+Install the following extensions:
+
+- [eslint](https://marketplace.visualstudio.com/items?itemName=dbaeumer.vscode-eslint)
+- [prettier](https://marketplace.visualstudio.com/items?itemName=esbenp.prettier-vscode)
+
+### Install dependencies
+
+Whenever dependencies in `package.json` are changed, run the following command:
+
+```sh
+pnpm install
+```
+
+To only install resolved dependencies in `package-lock.json`:
+
+```sh
+pnpm install --frozen-lockfile
+```
+
+## Other scripts for Development
+
+### Rebuild the project
+
+To incrementally build the project:
+
+```sh
+pnpm run build
+```
+
+### Fix code style and formatting issues
+
+(Using `lint` will also call `prettier` afterwards. For a pure `typescript-eslint` execution use `tslint`)
+
+```sh
+pnpm run lint
+```
+
+To automatically fix such issues:
+
+```sh
+pnpm run lint:fix
+```
