+++ kipper/target-js/README.md
@@ -1,0 +1,55 @@
+![](https://github.com/Luna-Klatzer/Kipper/raw/main/img/Kipper-Logo-with-head.png)
+
+# Kipper JavaScript Target - `@kipper/target-js`
+
+[![Version](https://img.shields.io/npm/v/@kipper/target-js?label=release&color=%23cd2620&logo=npm)](https://npmjs.org/package/@kipper/target-js)
+![](https://img.shields.io/badge/Coverage-78%25-5A7302.svg?style=flat&logoColor=white&color=blue&prefix=$coverage$)
+[![Issues](https://img.shields.io/github/issues/Luna-Klatzer/Kipper)](https://github.com/Luna-Klatzer/Kipper/issues)
+[![License](https://img.shields.io/github/license/Luna-Klatzer/Kipper?color=cyan)](https://github.com/Luna-Klatzer/Kipper/blob/main/LICENSE)
+[![Install size](https://packagephobia.com/badge?p=@kipper/target-js)](https://packagephobia.com/result?p=@kipper/target-js)
+[![Publish size](https://badgen.net/packagephobia/publish/@kipper/target-js)](https://packagephobia.com/result?p=@kipper/target-js)
+
+The JavaScript target for the Kipper Compiler.
+
+Kipper is a JavaScript-like strongly and strictly typed language with Python flavour. It aims to provide
+straightforward, simple, secure and type-safe coding with better efficiency and developer satisfaction! 🦊
+
+## Installation
+
+To install the whole Kipper package with its CLI, run the following command:
+
+```bash
+npm i @kipper/target-js
+```
+
+If you are using `pnpm` and `yarn`, use `pnpm i @kipper/target-js` or `yarn add @kipper/target-js`.
+
+## Kipper Docs
+
+Proper documentation for the Kipper language is available [here](https://luna-klatzer.github.io/Kipper/)!
+
+## Copyright and License
+
+![License](https://img.shields.io/github/license/Luna-Klatzer/Kipper?color=cyan)
+[![FOSSA Status](https://app.fossa.com/api/projects/git%2Bgithub.com%2FLuna-Klatzer%2FKipper.svg?type=shield)](https://app.fossa.com/projects/git%2Bgithub.com%2FLuna-Klatzer%2FKipper?ref=badge_shield)
+
+Copyright (C) 2021-2022 Luna Klatzer
+
+This program is free software: you can redistribute it and/or modify it under
+the terms of the GNU General Public License as published by the Free Software
+Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful, but WITHOUT ANY
+WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+PARTICULAR PURPOSE. See the GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License along with
+this program. If not, see <https://www.gnu.org/licenses/>.
+
+See the [LICENSE](https://raw.githubusercontent.com/Luna-Klatzer/Kipper/main/LICENSE)
+for information on terms & conditions for usage.
+
+### FOSSA License Report
+
+[![FOSSA Status](https://app.fossa.com/api/projects/git%2Bgithub.com%2FLuna-Klatzer%2FKipper.svg?type=large)](https://app.fossa.com/projects/git%2Bgithub.com%2FLuna-Klatzer%2FKipper?ref=badge_large)
