+++ kipper/target-js/src/built-in-generator.ts
@@ -3,56 +3,55 @@
  * functions.
  * @author Luna Klatzer
  * @copyright 2021-2022 Luna Klatzer
- * @since 0.8.0
+ * @since 0.10.0
  */
 import type { BuiltInFunction, TranslatedCodeLine } from "@kipper/core";
-import { getJavaScriptBuiltInIdentifier, JavaScriptTargetBuiltInGenerator } from "@kipper/target-js";
-import { getTypeScriptType } from "./tools";
+import { KipperTargetBuiltInGenerator } from "@kipper/core";
 
 /**
- * Generates the signature for the function based on the {@link funcSpec}, which can be used in an TypeScript env.
+ * Generates the signature for the function based on the {@link funcSpec}, which can be used in an JavaScript env.
  * @param funcSpec The function spec object containing the metadata of the function.
+ * @since 0.10.0
  */
-function getTSFunctionSignature(funcSpec: BuiltInFunction): {
-	type: string;
+function getJSFunctionSignature(funcSpec: BuiltInFunction): {
 	identifier: string;
-	args: Array<[string, string]>;
+	args: Array<string>;
 } {
-	// Translate the function signature into TypeScript
+	// Translate the function signature into JavaScript
 	const identifier: string = funcSpec.identifier;
-	const type: string = getTypeScriptType(funcSpec.returnType);
-	const args: Array<[string, string]> = funcSpec.args.map((arg) => [arg.identifier, getTypeScriptType(arg.type)]);
+	const args: Array<string> = funcSpec.args.map((arg) => arg.identifier);
 
-	return { type, identifier, args };
+	return { identifier, args };
 }
 
-function createTSFunctionSignature(signature: {
-	type: string;
-	identifier: string;
-	args: Array<[string, string]>;
-}): string {
-	const { type, identifier, args } = signature;
-	return `function ${identifier}(${args.map((arg) => `${arg[0]}: ${arg[1]}`).join(", ")}): ${type}`;
+/**
+ * Generates the JavaScript function signature, based on the {@link signature signature metadata}.
+ * @param signature The function signature metadata.
+ * @since 0.10.0
+ */
+function createJSFunctionSignature(signature: { identifier: string; args: Array<string> }): string {
+	const { identifier, args } = signature;
+	return `function ${identifier}(${args.join(", ")})`;
 }
 
 /**
  * The TypeScript target-specific built-ins generator for generating the code that allows for the use of built-in
  * functions.
- * @since 0.8.0
+ * @since 0.10.0
  */
-export class TypeScriptTargetBuiltInGenerator extends JavaScriptTargetBuiltInGenerator {
+export class JavaScriptTargetBuiltInGenerator extends KipperTargetBuiltInGenerator {
 	async numToStr(funcSpec: BuiltInFunction): Promise<Array<TranslatedCodeLine>> {
-		const signature = getTSFunctionSignature(funcSpec);
+		const signature = getJSFunctionSignature(funcSpec);
 		const convArgIdentifier = signature.args[0][0];
 
 		// Define the function signature and its body. We will simply use 'console.log(msg)' for printing out IO.
 		return [
 			[
-				getJavaScriptBuiltInIdentifier(signature.identifier),
+				`__kipper.${signature.identifier}`,
 				" ",
 				"=",
 				" ",
-				createTSFunctionSignature(signature),
+				createJSFunctionSignature(signature),
 				" ",
 				`{ return (${convArgIdentifier}).toString(); }`,
 				";",
@@ -61,17 +60,17 @@ export class TypeScriptTargetBuiltInGenerator extends JavaScriptTargetBuiltInGen
 	}
 
 	async strToNum(funcSpec: BuiltInFunction): Promise<Array<TranslatedCodeLine>> {
-		const signature = getTSFunctionSignature(funcSpec);
+		const signature = getJSFunctionSignature(funcSpec);
 		const convArgIdentifier = signature.args[0][0];
 
 		// Define the function signature and its body. We will simply use 'console.log(msg)' for printing out IO.
 		return [
 			[
-				getJavaScriptBuiltInIdentifier(signature.identifier),
+				`__kipper.${signature.identifier}`,
 				" ",
 				"=",
 				" ",
-				createTSFunctionSignature(signature),
+				createJSFunctionSignature(signature),
 				" ",
 				`{ return parseInt(${convArgIdentifier}); }`,
 				";",
@@ -80,17 +79,17 @@ export class TypeScriptTargetBuiltInGenerator extends JavaScriptTargetBuiltInGen
 	}
 
 	async boolToStr(funcSpec: BuiltInFunction): Promise<Array<TranslatedCodeLine>> {
-		const signature = getTSFunctionSignature(funcSpec);
+		const signature = getJSFunctionSignature(funcSpec);
 		const convArgIdentifier = signature.args[0][0];
 
 		// Define the function signature and its body. We will simply use 'console.log(msg)' for printing out IO.
 		return [
 			[
-				getJavaScriptBuiltInIdentifier(signature.identifier),
+				`__kipper.${signature.identifier}`,
 				" ",
 				"=",
 				" ",
-				createTSFunctionSignature(signature),
+				createJSFunctionSignature(signature),
 				" ",
 				`{ return \`\${${convArgIdentifier}}\`; }`,
 				";",
@@ -99,17 +98,17 @@ export class TypeScriptTargetBuiltInGenerator extends JavaScriptTargetBuiltInGen
 	}
 
 	async boolToNum(funcSpec: BuiltInFunction): Promise<Array<TranslatedCodeLine>> {
-		const signature = getTSFunctionSignature(funcSpec);
+		const signature = getJSFunctionSignature(funcSpec);
 		const convArgIdentifier = signature.args[0][0];
 
 		// Define the function signature and its body. We will simply use 'console.log(msg)' for printing out IO.
 		return [
 			[
-				getJavaScriptBuiltInIdentifier(signature.identifier),
+				`__kipper.${signature.identifier}`,
 				" ",
 				"=",
 				" ",
-				createTSFunctionSignature(signature),
+				createJSFunctionSignature(signature),
 				" ",
 				`{ return ${convArgIdentifier} ? 1 : 0; }`,
 				";",
@@ -118,17 +117,17 @@ export class TypeScriptTargetBuiltInGenerator extends JavaScriptTargetBuiltInGen
 	}
 
 	async print(funcSpec: BuiltInFunction): Promise<Array<TranslatedCodeLine>> {
-		const signature = getTSFunctionSignature(funcSpec);
+		const signature = getJSFunctionSignature(funcSpec);
 		const printArgIdentifier = signature.args[0][0];
 
 		// Define the function signature and its body. We will simply use 'console.log(msg)' for printing out IO.
 		return [
 			[
-				getJavaScriptBuiltInIdentifier(signature.identifier),
+				`__kipper.${signature.identifier}`,
 				" ",
 				"=",
 				" ",
-				createTSFunctionSignature(signature),
+				createJSFunctionSignature(signature),
 				" ",
 				`{ console.log(${printArgIdentifier}); }`,
 				";",
