+++ kipper/target-js/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+		"outDir": "./lib",
+		"rootDir": "./src",
+		"noImplicitAny": true,
+		"strict": true,
+		"target": "es2020",
+		"lib": [
+			"ES7"
+		]
+		// ES7 -> ES2016
+  },
+  "include": [
+    "./src/**/*"
+  ],
+	"exclude": [
+		"node_modules",
+		"lib",
+		"dist",
+		"**/*.spec.ts"
+	],
+	"files": [],
+}
