+++ package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@kipper/cli": "workspace:~",
     "@kipper/core": "workspace:~",
+		"@kipper/target-js": "workspace:~",
 		"@kipper/target-ts": "workspace:~",
     "antlr4ts": "^0.5.0-alpha.4",
     "tslib": "~2.4.0",
